changed alarm space of 1 to 2 so it has a minute between alarms

+++ app/src/main/java/com/example/myfirstapp/intro.java
@@ -103,7 +103,7 @@ public class intro extends AppCompatActivity {
         findViewById(R.id.mko).setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                alarmSpace = new Integer(1);
+                alarmSpace = new Integer(2);
                 me.setText(alarmSpace.toString());
             }
         });
